requirements.txt
instructor
wandb
pandas
python-dotenv